refacto plugin logger and rename repo file
