+++ lib/infrastructure/CatalogQueryRepository.ts
